+++ puppetreport/collector_windows.go
@@ -16,9 +16,12 @@
 
 package puppetreport
 
-func (c Collector) reportPath() string {
+// DefaultReportPath is the default location of the puppet agent last run report file on windows.
+const DefaultReportPath = "C:/ProgramData/PuppetLabs/puppet/cache/state/last_run_report.yaml"
+
+func (c *Collector) reportPath() string {
 	if c.ReportPath != "" {
 		return c.ReportPath
 	}
-	return "C:/ProgramData/PuppetLabs/puppet/cache/state/last_run_report.yaml"
+	return DefaultReportPath
 }